+++ ocrd_tesserocr/recognize.py
@@ -92,7 +92,9 @@ class TesserocrRecognize(Processor):
         
         Set up Tesseract to recognise each segment's image (either from
         AlternativeImage or cropping the bounding box rectangle and masking
-        it from the polygon outline) with the appropriate mode and ``model``.
+        it from the polygon outline) with the appropriate segmentation mode
+        and ``model``. (If no ``model`` is given, only layout analysis will
+        be performed.)
         
         Next, if there still is a gap between the current level in the PAGE hierarchy
         and the requested ``textequiv_level``, then iterate down the result hierarchy,
@@ -100,7 +102,8 @@ class TesserocrRecognize(Processor):
         text line order, reading direction and orientation at the region/table level).
         
         Then, at ``textequiv_level``, remove any existing TextEquiv, unless
-        ``overwrite_text`` is false, and add text and confidence results .
+        ``overwrite_text`` is false, and add text and confidence results, unless
+        ``model`` is empty.
         
         The special value ``textequiv_level=none`` behaves like ``glyph``,
         except that no actual text recognition will be performed, only
@@ -121,28 +124,33 @@ class TesserocrRecognize(Processor):
         In other words:
         - If ``segmentation_level=region``, then segment the page into regions
           (unless ``overwrite_segments=false``), else iterate existing regions.
-        - If ``textequiv_level=region``, then recognize text in the region,
-          annotate it, and continue with the next region. Otherwise...
+        - If ``textequiv_level=region``, then unless ``model`` is empty,
+          recognize text in the region and annotate it. Regardless, continue
+          with the next region. Otherwise...
         - If ``segmentation_level=cell`` or higher,
           then segment table regions into text regions (i.e. cells)
           (unless ``overwrite_segments=false``), else iterate existing cells.
-        - If ``textequiv_level=cell``, then recognize text in the cell,
-          annotate it, and continue with the next cell. Otherwise...
+        - If ``textequiv_level=cell``, then unless ``model`` is empty,
+          recognize text in the cell and annotate it. Regardless, continue
+          with the next cell. Otherwise...
         - If ``segmentation_level=line`` or higher,
           then segment text regions into text lines
           (unless ``overwrite_segments=false``), else iterate existing text lines.
-        - If ``textequiv_level=line``, then recognize text in the text lines,
-          annotate it, and continue with the next line. Otherwise...
+        - If ``textequiv_level=line``, then unless ``model`` is empty,
+          recognize text in the text lines and annotate it. Regardless, continue
+          with the next line. Otherwise...
         - If ``segmentation_level=word`` or higher,
           then segment text lines into words
           (unless ``overwrite_segments=false``), else iterate existing words.
-        - If ``textequiv_level=word``, then recognize text in the words,
-          annotate it, and continue with the next word. Otherwise...
+        - If ``textequiv_level=word``, then unless ``model`` is empty,
+          recognize text in the words and annotate it. Regardless, continue
+          with the next word. Otherwise...
         - If ``segmentation_level=glyph`` or higher,
           then segment words into glyphs
           (unless ``overwrite_segments=false``), else iterate existing glyphs.
-        - If ``textequiv_level=glyph``, then recognize text in the glyphs and
-          continue with the next glyph. Otherwise...
+        - If ``textequiv_level=glyph``, then unless ``model`` is empty,
+          recognize text in the glyphs and annotate it. Regardless, continue
+          with the next glyph. Otherwise...
         - (i.e. ``none``) annotate no text and be done.
         
         Note that ``cell`` is an _optional_ level that is only relevant for
@@ -156,7 +164,8 @@ class TesserocrRecognize(Processor):
         Thus, ``segmentation_level`` is the entry point level for layout analysis,
         and setting it to ``none`` makes this processor behave as recognition-only.
         Whereas ``textequiv_level`` selects the exit point level for segmentation,
-        and setting it to ``none`` makes this processor behave as segmentation-only.
+        and setting it to ``none`` makes this processor behave as segmentation-only,
+        as does omitting ``model``.
         
         All segments above ``segmentation_level`` must already exist, and
         no segments below ``textequiv_level`` will be newly created.
@@ -196,15 +205,15 @@ class TesserocrRecognize(Processor):
             for sub_model in model.split('+'):
                 if sub_model not in get_languages()[1]:
                     raise Exception("configured model " + sub_model + " is not installed")
+                self.logger.info("Using model '%s' in %s for recognition at the %s level",
+                                 model, get_languages()[0], outlevel)
         
         with PyTessBaseAPI(path=TESSDATA_PREFIX, lang=model) as tessapi:
-            self.logger.info("Using model '%s' in %s for recognition at the %s level",
-                             model, get_languages()[0], outlevel)
             if outlevel == 'glyph':
                 # populate GetChoiceIterator() with LSTM models, too:
                 tessapi.SetVariable("lstm_choice_mode", "2") # aggregate symbols
                 tessapi.SetVariable("lstm_choice_iterations", "15") # squeeze out more best paths
-            if outlevel != 'none':
+            if outlevel != 'none' and self.parameter.get('model', ''):
                 # TODO: maybe warn/raise when illegal combinations or characters not in the model unicharset?
                 if self.parameter['char_whitelist']:
                     tessapi.SetVariable("tessedit_char_whitelist", self.parameter['char_whitelist'])
@@ -313,7 +322,7 @@ class TesserocrRecognize(Processor):
                     page.set_ReadingOrder(None)
                     # prepare Tesseract
                     if self.parameter['find_tables']:
-                        if outlevel == 'region':
+                        if outlevel == 'region' and self.parameter.get('model', ''):
                             raise Exception("When segmentation_level is region and find_tables is enabled, textequiv_level must be at least cell, because text results cannot be annotated on tables directly.")
                         tessapi.SetVariable("textord_tabfind_find_tables", "1") # (default)
                         # this should yield additional blocks within the table blocks
@@ -329,7 +338,7 @@ class TesserocrRecognize(Processor):
                     tessapi.SetPageSegMode(PSM.SPARSE_TEXT
                                            if self.parameter['sparse_text']
                                            else PSM.AUTO)
-                    if outlevel == 'none':
+                    if outlevel == 'none' or not self.parameter.get('model', ''):
                         self.logger.debug("Detecting regions in page '%s'", page_id)
                         tessapi.AnalyseLayout()
                     else:
@@ -359,7 +368,7 @@ class TesserocrRecognize(Processor):
                 
                 # post-processing
                 # bottom-up text concatenation
-                if outlevel != 'none':
+                if outlevel != 'none' and self.parameter.get('model', ''):
                     page_update_higher_textequiv_levels(outlevel, pcgts)
                 # bottom-up polygonal outline projection
                 # if inlevel != 'none' and self.parameter['shrink_polygons']:
@@ -482,13 +491,13 @@ class TesserocrRecognize(Processor):
                     region.set_type(TextTypeSimpleType.CAPTION)
                 page.add_TextRegion(region)
                 og.add_RegionRefIndexed(RegionRefIndexedType(regionRef=ID, index=index))
-                if self.parameter['textequiv_level'] == 'region':
+                if self.parameter['textequiv_level'] != 'region':
+                    self._process_lines_in_region(it, region, page_coords)
+                elif self.parameter.get('model', ''):
                     region.add_TextEquiv(TextEquivType(
                         Unicode=it.GetUTF8Text(RIL.BLOCK).rstrip("\n\f"),
                         # iterator scores are arithmetic averages, too
                         conf=it.Confidence(RIL.BLOCK)/100.0))
-                else:
-                    self._process_lines_in_region(it, region, page_coords)
             elif block_type in [PT.FLOWING_IMAGE,
                                 PT.HEADING_IMAGE,
                                 PT.PULLOUT_IMAGE]:
@@ -577,14 +586,14 @@ class TesserocrRecognize(Processor):
             self._add_orientation(it, cell, page_coords)
             if rogroup:
                 rogroup.add_RegionRefIndexed(RegionRefIndexedType(regionRef=ID, index=index))
-            if self.parameter['textequiv_level'] == 'cell':
+            if self.parameter['textequiv_level'] != 'cell':
+                self._process_lines_in_region(it, cell, page_coords, parent_ril=ril)
+            elif self.parameter.get('model', ''):
                 cell.add_TextEquiv(TextEquivType(
                     Unicode=it.GetUTF8Text(ril).rstrip("\n\f"),
                     # iterator scores are arithmetic averages, too
                     conf=it.Confidence(ril)/100.0))
-            else:
-                self._process_lines_in_region(it, cell, page_coords, parent_ril=ril)
-        
+
     def _process_lines_in_region(self, result_it, region, page_coords, parent_ril=RIL.BLOCK):
         if self.parameter['sparse_text']:
             it = result_it
@@ -592,14 +601,14 @@ class TesserocrRecognize(Processor):
             line = TextLineType(id=region.id + '_line',
                                 Coords=region.get_Coords())
             region.add_TextLine(line)
-            if self.parameter['textequiv_level'] == 'line':
+            if self.parameter['textequiv_level'] != 'line':
+                self._process_words_in_line(it, line, page_coords)
+            elif self.parameter.get('model', ''):
                 # todo: consider BlankBeforeWord, SetLineSeparator
                 line.add_TextEquiv(TextEquivType(
                     Unicode=it.GetUTF8Text(RIL.TEXTLINE).rstrip("\n\f"),
                     # iterator scores are arithmetic averages, too
                     conf=it.Confidence(RIL.TEXTLINE)/100.0))
-            else:
-                self._process_words_in_line(it, line, page_coords)
             return
         for index, it in enumerate(iterate_level(result_it, RIL.TEXTLINE, parent=parent_ril)):
             bbox = it.BoundingBox(RIL.TEXTLINE, padding=self.parameter['padding'])
@@ -623,15 +632,15 @@ class TesserocrRecognize(Processor):
             self.logger.info("Detected line '%s': %s", ID, points)
             line = TextLineType(id=ID, Coords=coords)
             region.add_TextLine(line)
-            if self.parameter['textequiv_level'] == 'line':
+            if self.parameter['textequiv_level'] != 'line':
+                self._process_words_in_line(it, line, page_coords)
+            elif self.parameter.get('model', ''):
                 # todo: consider BlankBeforeWord, SetLineSeparator
                 line.add_TextEquiv(TextEquivType(
                     Unicode=it.GetUTF8Text(RIL.TEXTLINE).rstrip("\n\f"),
                     # iterator scores are arithmetic averages, too
                     conf=it.Confidence(RIL.TEXTLINE)/100.0))
-            else:
-                self._process_words_in_line(it, line, page_coords)
-    
+
     def _process_words_in_line(self, result_it, line, coords):
         for index, it in enumerate(iterate_level(result_it, RIL.WORD)):
             bbox = it.BoundingBox(RIL.WORD, padding=self.parameter['padding'])
@@ -658,14 +667,14 @@ class TesserocrRecognize(Processor):
             self.logger.debug("Detected word '%s': %s", ID, points)
             word = WordType(id=ID, Coords=CoordsType(points=points))
             line.add_Word(word)
-            if self.parameter['textequiv_level'] == 'word':
+            if self.parameter['textequiv_level'] != 'word':
+                self._process_glyphs_in_word(it, word, coords)
+            elif self.parameter.get('model', ''):
                 word.add_TextEquiv(TextEquivType(
                     Unicode=it.GetUTF8Text(RIL.WORD),
                     # iterator scores are arithmetic averages, too
                     conf=it.Confidence(RIL.WORD)/100.0))
-            else:
-                self._process_glyphs_in_word(it, word, coords)
-    
+
     def _process_glyphs_in_word(self, result_it, word, coords):
         for index, it in enumerate(iterate_level(result_it, RIL.SYMBOL)):
             bbox = it.BoundingBox(RIL.SYMBOL, padding=self.parameter['padding'])
@@ -682,7 +691,9 @@ class TesserocrRecognize(Processor):
             #self.logger.debug("Detected glyph '%s': %s", ID, points)
             glyph = GlyphType(id=ID, Coords=CoordsType(points))
             word.add_Glyph(glyph)
-            if self.parameter['textequiv_level'] == 'glyph':
+            if self.parameter['textequiv_level'] != 'glyph':
+                pass
+            elif self.parameter.get('model', ''):
                 # glyph_text = it.GetUTF8Text(RIL.SYMBOL) # equals first choice?
                 glyph_conf = it.Confidence(RIL.SYMBOL)/100 # equals first choice?
                 #self.logger.debug('best glyph: "%s" [%f]', glyph_text, glyph_conf)
@@ -772,7 +783,7 @@ class TesserocrRecognize(Processor):
                 tessapi.SetImage(table_image)
             tessapi.SetPageSegMode(PSM.SPARSE_TEXT) # retrieve "cells"
             # TODO: we should XY-cut the sparse cells in regroup them into consistent cells
-            if self.parameter['textequiv_level'] == 'none':
+            if self.parameter['textequiv_level'] == 'none' or not self.parameter.get('model', ''):
                 self.logger.debug("Detecting cells in table '%s'", table.id)
                 tessapi.AnalyseLayout()
             else:
@@ -781,13 +792,15 @@ class TesserocrRecognize(Processor):
             self._process_cells_in_table(tessapi.GetIterator(), table, roelem, table_coords)
     
     def _process_existing_regions(self, tessapi, regions, page_image, page_coords):
+        if self.parameter['textequiv_level'] in ['region', 'cell'] and not self.parameter.get('model', ''):
+            return
         for region in regions:
             region_image, region_coords = self.workspace.image_from_segment(
                 region, page_image, page_coords)
             if not region_image.width or not region_image.height:
                 self.logger.warning("Skipping text region '%s' with zero size", region.id)
                 continue
-            if (self.parameter['textequiv_level'] != 'region' and
+            if (self.parameter['textequiv_level'] not in ['region', 'cell'] and
                 self.parameter['segmentation_level'] != 'line'):
                 pass # image not used here
             elif self.parameter['padding']:
@@ -817,7 +830,7 @@ class TesserocrRecognize(Processor):
                 if textlines:
                     self.logger.info('Removing existing text lines in region %s', region.id)
                 region.set_TextLine([])
-                if self.parameter['textequiv_level'] == 'none':
+                if self.parameter['textequiv_level'] == 'none' or not self.parameter.get('model', ''):
                     self.logger.debug("Detecting lines in region '%s'", region.id)
                     tessapi.AnalyseLayout()
                 else:
@@ -831,6 +844,8 @@ class TesserocrRecognize(Processor):
                                     region.id)
 
     def _process_existing_lines(self, tessapi, textlines, region_image, region_coords):
+        if self.parameter['textequiv_level'] == 'line' and not self.parameter.get('model', ''):
+            return
         for line in textlines:
             line_image, line_coords = self.workspace.image_from_segment(
                 line, region_image, region_coords)
@@ -869,7 +884,7 @@ class TesserocrRecognize(Processor):
                 if words:
                     self.logger.info('Removing existing words in line %s', line.id)
                 line.set_Word([])
-                if self.parameter['textequiv_level'] == 'none':
+                if self.parameter['textequiv_level'] == 'none' or not self.parameter.get('model', ''):
                     self.logger.debug("Detecting words in line '%s'", line.id)
                     tessapi.AnalyseLayout()
                 else:
@@ -886,6 +901,8 @@ class TesserocrRecognize(Processor):
                                     line.id)
 
     def _process_existing_words(self, tessapi, words, line_image, line_coords):
+        if self.parameter['textequiv_level'] == 'word' and not self.parameter.get('model', ''):
+            return
         for word in words:
             word_image, word_coords = self.workspace.image_from_segment(
                 word, line_image, line_coords)
@@ -919,7 +936,7 @@ class TesserocrRecognize(Processor):
                 if glyphs:
                     self.logger.info('Removing existing glyphs in word %s', word.id)
                 word.set_Glyph([])
-                if self.parameter['textequiv_level'] == 'none':
+                if self.parameter['textequiv_level'] == 'none' or not self.parameter.get('model', ''):
                     self.logger.debug("Detecting glyphs in word '%s'", word.id)
                     tessapi.AnalyseLayout()
                 else:
@@ -936,6 +953,8 @@ class TesserocrRecognize(Processor):
                                     word.id)
 
     def _process_existing_glyphs(self, tessapi, glyphs, word_image, word_xywh):
+        if not self.parameter.get('model', ''):
+            return
         for glyph in glyphs:
             glyph_image, _ = self.workspace.image_from_segment(
                 glyph, word_image, word_xywh)