+++ ocrd_tesserocr/segment_table.py
@@ -24,7 +24,7 @@ class TesserocrSegmentTable(Processor):
             recognize_kwargs['parameter']['overwrite_segments'] = self.parameter['overwrite_cells']
             del recognize_kwargs['parameter']['overwrite_regions']
             recognize_kwargs['parameter']['segmentation_level'] = "cell"
-            recognize_kwargs['parameter']['textequiv_level'] = "none"
+            recognize_kwargs['parameter']['textequiv_level'] = "cell"
             self.recognizer = TesserocrRecognize(self.workspace, **recognize_kwargs)
             self.recognizer.logger = getLogger('processor.TesserocrSegmentTable')