+++ ocrd_tesserocr/segment_word.py
@@ -24,7 +24,7 @@ class TesserocrSegmentWord(Processor):
             recognize_kwargs['parameter']['overwrite_segments'] = self.parameter['overwrite_words']
             del recognize_kwargs['parameter']['overwrite_words']
             recognize_kwargs['parameter']['segmentation_level'] = "word"
-            recognize_kwargs['parameter']['textequiv_level'] = "none"
+            recognize_kwargs['parameter']['textequiv_level'] = "word"
             self.recognizer = TesserocrRecognize(self.workspace, **recognize_kwargs)
             self.recognizer.logger = getLogger('processor.TesserocrSegmentWord')